Revert ae2ced5 and fix tests
I accidentally copy-pasted an URL with an `access_token` query parameter
for testing the GET verb on the `/directory/room/:room_alias` endpoint
in 6cd596d. This was never used in the implementation of the request
because according to the spec [1], GET on this endpoint does not require
authentication. This commit reverts the addition of the
`AccessTokenAuth` middleware done in ae2ced5 and removes the
`access_token` parameter from the tests.

[1] http://matrix.org/docs/spec/client_server/r0.1.0.html#get-matrix-client-r0-directory-room-roomalias