Require authentication for deleting room aliases
This adds an authentication middleware to DELETE
/directory/room/:room_alias, as mandated by the spec, and updates the
test accordingly. The test now actually checks the status code of the
response to the DELETE request. The mapping of `DieselError::NotFound`
to the appropiate `APIError` has been removed as a DELETE will finish
successfully even if nothing has been deleted actually. Furthermore, the
`delete` helper method in `test.rs` has been modified to not require a
body.

Marks the endpoint as "under construction" in STATUS.md, as we still
need to check if the requesting user has enough power to delete a room
alias.