Implement PUT /directory/room/:room_alias
This implementation does not return the correct error code (409) if the
room alias that should be created already exists. That is because
currently with Diesel there is no DBMS-independent way to tell that an
insert failed because there exists a row with the same primary key as
the row we are trying to insert.

Closes #19.