Ensure event creation requests are idempotent using transaction IDs
The url path of the endpoint along with the access token of the user
were used as the primary key for the transactions table. The path encapsulates
all the necessary information (room_id, event_type etc) to identify the
correct transaction without the need to store them explicitly.

Storing them explicitly would mean a more complex schema and logic
to identify the correct transaction because other endpoints with different
identifiers will also use transactions.